cdk/requirements.txt
# DTL-Global Platform — CDK app dependencies (Python 3.12+)
aws-cdk-lib>=2.100.0
constructs>=10.0.0
